+++ frontend/openchat-shared/src/domain/worker.ts
@@ -1046,7 +1046,7 @@ type InviteUsers = {
 };
 
 type RemoveSub = {
-    subscription: PushSubscriptionJSON;
+    endpoint: string;
     kind: "removeSubscription";
 };
 
@@ -1057,7 +1057,6 @@ type PushSub = {
 
 type SubscriptionExists = {
     endpoint: string;
-    p256dh_key: string;
     kind: "subscriptionExists";
 };